Compiler warnings

+++ src/wp43s/graphs.c
@@ -204,7 +204,7 @@ void graph_setupmemory(void) {
   } else
   {
   #ifdef PC_BUILD
-    printf("^^@@ Two arrays of %ld bytes each created, i.e. %ld blocks total\n",LIM * sizeof(graphtype), 2 * LIM * sizeof(graphtype) / 4);
+    printf("^^@@ Two arrays of %llu bytes each created, i.e. %llu blocks total\n",(long long unsigned int) (LIM * sizeof(graphtype)), (long long unsigned int)(2 * LIM * sizeof(graphtype) / 4));
   #endif
   }
   
@@ -213,8 +213,7 @@ void graph_setupmemory(void) {
       gr_x[i] = 0;
       gr_y[i] = 0; 
     }
-  }   
-  
+  }
 }
